+++ resources/views/collections/form.blade.php
@@ -87,9 +87,15 @@
 <div class="col-md-6">
 	<div class="form-group">
 		{!! Form::label('main_image', 'Obrázok') !!}
-		{!! Form::file('main_image') !!}
+		{!! Form::file('main_image', array('class' => 'form-control', 'placeholder' => '#666666'))!!}
 		<p>obrazok bude automaticky zmenseny na sirku 1400px</p>
 		<p>šírka min: 1400px<br>formát: JPG (vysoka kompresia ~50-60%)</p>
+		@if (isset($collection) && $collection->getHeaderImage())
+			<div class="primary-image">
+				<b>Aktuálny obrázok:</b>
+				<img src="{{ $collection->getHeaderImage() }}" class="img-responsive">
+			</div>
+		@endif
 	</div>
 </div>
 @if (Entrust::hasRole('admin'))