+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file[^1].
 ### Fixed
 - open importing csv files with read permissions only
 - ordering of catalogue by the view count
+- indexing of item's work types 
 
 ## [2.6.3] - 2020-05-12
 ### Fixed