+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 ### Fixed
+- preview main / header image within article and collection admin editor
+- prevent storing of empty alternative locales for articles / collections
+- small locale fixes
+- absolute urls for og:image
 - open importing csv files with read permissions only
 - ordering of catalogue by the view count