+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file[^1].
 - open importing csv files with read permissions only
 - ordering of catalogue by the view count
 
+### Changed
+- COVID-19 alert for printed reproductions
+- align and layout in order section
+
 ## [2.6.3] - 2020-05-12
 ### Fixed
 - creating order with empty address